worked on adding songs to playlist

+++ Musify Desktop App/Musify Desktop App/Model/PlaylistSong.cs
@@ -6,7 +6,7 @@ namespace Musify_Desktop_App.Model
     {
         public virtual Playlist Playlist { get; set; }
         public Song Song { get; set; }
-        public long SongId { get; set; }
+        public int SongId { get; set; }
         public int Number { get; set; }
         public DateTime DateAdded { get; set; }
     }
